+++ libsq/core/ocache/ocache.go
@@ -1,8 +1,8 @@
-// Package oncecache provides a strongly-typed, concurrency-safe, in-memory,
-// on-demand cache that ensures that a given cache entry is populated only once.
-// It also provides a callback mechanism that can be used for linked cache
-// propagation, logging, or metrics.
-package oncecache
+// Package ocache provides a strongly-typed, concurrency-safe, in-memory,
+// on-demand object cache that ensures that a given cache entry is populated
+// only once. It also provides a callback mechanism that can be used for linked
+// cache propagation, logging, or metrics.
+package ocache
 
 import (
 	"context"
